refactor: move pdf route models
