+++ convert/src/badgerdoc_format/badgerdoc_format.py
@@ -25,7 +25,6 @@ class BadgerdocFormat:
     def export_tokens(self, path: Path) -> None:
         if self.tokens_page:
             path.write_text(self.tokens_page.json(by_alias=True))
-
     def export_annotations(self, path: Path):
         if self.badgerdoc_annotation:
             path.write_text(self.badgerdoc_annotation.json(indent=4))
